# 3.x: TokenAwarePolicy fix bad perf of ReplicaOrdering.RANDOM