@mykaul 2025-01-27
Do we have a similar optimization potential with 4.x, or is this 3.x-specific? (did not look at the code yet)