@Bouncheck 2025-01-27
> Do we have a similar optimization potential with 4.x, or is this 3.x-specific? (did not look at the code yet)

Looks like 4.x already uses `ThreadLocalRandom`. There are two uses of `Collections.shuffle` without it but those are related to reconnection, so that won't have much impact.