@roydahan 2025-01-27
Thanks @dkropachev, great work identifying the root cause.

@juliayakovlev this should solve the issue you have with c-s 3.17.0, we will need a new c-s release.